+++ database/migrations/2022_05_17_091555_create_barangs_table.php
@@ -17,8 +17,8 @@ class CreateBarangsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('kategori');
-            $table->decimal('harga');
-            $table->decimal('stock');
+            $table->integer('harga');
+            $table->integer('stock');
             $table->timestamps();
             $table->softDeletes();
         });
